+++ lib/shrine/plugins/processing.rb
@@ -1,10 +1,10 @@
 class Shrine
   module Plugins
     # The process plugin allows you to declaratively define
-    # file processing for specified phases, allowing you to transform
+    # file processing for specified actions, allowing you to transform
     #
     #     def process(io, context)
-    #       if context[:phase] == :store
+    #       if context[:action] == :store
     #         # ...
     #       end
     #     end
@@ -15,7 +15,7 @@ class Shrine
     #       # ...
     #     end
     #
-    # The declarations are additive and inherited, so for the same phase you
+    # The declarations are additive and inherited, so for the same action you
     # can declare multiple blocks, and they will be performed in the same order,
     # where output from previous will be input to next. You can return `nil`
     # in any block to signal that no processing was performed and that the
@@ -26,15 +26,15 @@ class Shrine
       end
 
       module ClassMethods
-        def process(phase, &block)
-          opts[:processing][phase] ||= []
-          opts[:processing][phase] << block
+        def process(action, &block)
+          opts[:processing][action] ||= []
+          opts[:processing][action] << block
         end
       end
 
       module InstanceMethods
         def process(io, context = {})
-          pipeline = opts[:processing][context[:phase]] || []
+          pipeline = opts[:processing][context[:action]] || []
 
           result = pipeline.inject(io) do |input, processing|
             instance_exec(input, context, &processing) || input