+++ lib/shrine/plugins/versions.rb
@@ -210,7 +210,7 @@ class Shrine
 
         def assign_cached(value)
           cached_file = uploaded_file(value)
-          warn "Generating versions in the :cache phase is deprecated and will be forbidden in Shrine 3." if cached_file.is_a?(Hash)
+          warn "Generating versions in the :cache action is deprecated and will be forbidden in Shrine 3." if cached_file.is_a?(Hash)
           super(cached_file)
         end
       end