+++ test/plugin/processing_test.rb
@@ -7,37 +7,37 @@ describe Shrine::Plugins::Processing do
   end
 
   it "executes defined processing" do
-    @uploader.class.process(:phase) { |io, context| FakeIO.new(io.read.reverse) }
-    uploaded_file = @uploader.upload(fakeio("file"), phase: :phase)
+    @uploader.class.process(:foo) { |io, context| FakeIO.new(io.read.reverse) }
+    uploaded_file = @uploader.upload(fakeio("file"), action: :foo)
     assert_equal "elif", uploaded_file.read
   end
 
   it "executes in context of uploader, and passes right variables" do
-    @uploader.class.process(:phase) do |io, context|
+    @uploader.class.process(:foo) do |io, context|
       raise unless self.is_a?(Shrine)
       raise unless io.respond_to?(:read)
-      raise unless context.is_a?(Hash) && context.key?(:phase)
+      raise unless context.is_a?(Hash) && context.key?(:action)
       FakeIO.new(io.read.reverse)
     end
-    @uploader.upload(fakeio("file"), phase: :phase)
+    @uploader.upload(fakeio("file"), action: :foo)
   end
 
   it "executes all defined blocks where output of previous is input to next" do
-    @uploader.class.process(:phase) { |io, context| FakeIO.new("changed") }
-    @uploader.class.process(:phase) { |io, context| FakeIO.new(io.read.reverse) }
-    uploaded_file = @uploader.upload(fakeio("file"), phase: :phase)
+    @uploader.class.process(:foo) { |io, context| FakeIO.new("changed") }
+    @uploader.class.process(:foo) { |io, context| FakeIO.new(io.read.reverse) }
+    uploaded_file = @uploader.upload(fakeio("file"), action: :foo)
     assert_equal "degnahc", uploaded_file.read
   end
 
   it "allows blocks to return nil" do
-    @uploader.class.process(:phase) { |io, context| nil }
-    @uploader.class.process(:phase) { |io, context| FakeIO.new(io.read.reverse) }
-    uploaded_file = @uploader.upload(fakeio("file"), phase: :phase)
+    @uploader.class.process(:foo) { |io, context| nil }
+    @uploader.class.process(:foo) { |io, context| FakeIO.new(io.read.reverse) }
+    uploaded_file = @uploader.upload(fakeio("file"), action: :foo)
     assert_equal "elif", uploaded_file.read
   end
 
   it "executes defined blocks only if phases match" do
-    @uploader.class.process(:phase) { |io, context| FakeIO.new(io.read.reverse) }
+    @uploader.class.process(:foo) { |io, context| FakeIO.new(io.read.reverse) }
     uploaded_file = @uploader.upload(fakeio("file"))
     assert_equal "file", uploaded_file.read
   end